+++ package.json
@@ -45,7 +45,7 @@
     "react-modal-promise": "^1.0.2",
     "react-redux": "^7.1.3",
     "react-refresh": "^0.11.0",
-    "react-router-dom": "^6.2.2",
+    "react-router-dom": "6.2.2",
     "react-text-mask": "^5.4.3",
     "react-transition-group": "^4.3.0",
     "redux": "^4.0.5",
