+++ src/elements/RegisterArtifactModalForm/RegisterArtifactModalForm.js
@@ -17,39 +17,15 @@ illegal under applicable law, and the grant of the foregoing license
 under the Apache 2.0 license is conditioned upon your compliance with
 such restriction.
 */
-import React, { useEffect, useMemo, useState } from 'react'
+import React, { useMemo } from 'react'
 import PropTypes from 'prop-types'
 
-import { FormInput, FormSelect, FormTextarea, FormCombobox } from 'igz-controls/components'
+import TargetPath from '../../common/TargetPath/TargetPath'
 
+import { FormInput, FormSelect, FormTextarea } from 'igz-controls/components'
 import { getValidationRules } from 'igz-controls/utils/validation.util'
-import {
-  dataInputInitialState,
-  pathTips,
-  pathPlaceholders,
-  handleStoreInputPathChange,
-  comboboxSelectList,
-  isPathInputInvalid,
-  generateProjectsList,
-  generateComboboxMatchesList,
-  generateArtifactsList,
-  generateArtifactsReferencesList
-} from '../../utils/combobox'
-import { MLRUN_STORAGE_INPUT_PATH_SCHEME } from '../../constants'
-import { isNil, pick, uniqBy } from 'lodash'
-import { useDispatch } from 'react-redux'
-import { useParams } from 'react-router-dom'
-import artifactsAction from '../../actions/artifacts'
-import featureStoreActions from '../../actions/featureStore'
-import { OnChange } from 'react-final-form-listeners'
-import projectAction from '../../actions/projects'
-import { getFeatureReference } from '../../utils/resources'
 
 const RegisterArtifactModalForm = ({ formState, showType, messageByKind, setFieldState }) => {
-  const [dataInputState, setDataInputState] = useState(dataInputInitialState)
-  const params = useParams()
-  const dispatch = useDispatch()
-
   const kindOptions = useMemo(
     () => [
       {
@@ -71,179 +47,6 @@ const RegisterArtifactModalForm = ({ formState, showType, messageByKind, setFiel
     ],
     []
   )
-
-  const handleOnChange = (selectValue, inputValue) => {
-    if (isNil(inputValue)) {
-      setFieldState('formState.values.target_path.path', { modified: false })
-    }
-
-    if (selectValue === MLRUN_STORAGE_INPUT_PATH_SCHEME && !isNil(inputValue)) {
-      handleStoreInputPathChange(dataInputState, setDataInputState, inputValue)
-    }
-  }
-
-  const validatePath = allValues => {
-    const { pathType, value } = pick(allValues.target_path.fieldInfo, ['pathType', 'value'])
-
-    return isPathInputInvalid(pathType, value)
-  }
-
-  useEffect(() => {
-    if (dataInputState.inputStorePathTypeEntered && dataInputState.projects.length === 0) {
-      dispatch(projectAction.fetchProjectsNames()).then(result => {
-        setDataInputState(prev => ({
-          ...prev,
-          projects: generateProjectsList(result, params.projectName)
-        }))
-      })
-    }
-  }, [
-    dataInputState.inputStorePathTypeEntered,
-    dataInputState.projects.length,
-    dispatch,
-    params.projectName,
-    setDataInputState
-  ])
-
-  useEffect(() => {
-    if (formState.values.target_path.fieldInfo.pathType === MLRUN_STORAGE_INPUT_PATH_SCHEME) {
-      setDataInputState(prev => ({
-        ...prev,
-        comboboxMatches: generateComboboxMatchesList(
-          dataInputState.artifacts,
-          dataInputState.artifactsReferences,
-          dataInputState.featureVectors,
-          dataInputState.featureVectorsReferences,
-          dataInputState.inputProjectItemPathEntered,
-          dataInputState.inputProjectItemReferencePathEntered,
-          dataInputState.inputProjectPathEntered,
-          dataInputState.inputStorePathTypeEntered,
-          dataInputState.project,
-          dataInputState.projectItem,
-          dataInputState.projectItemReference,
-          dataInputState.projects,
-          dataInputState.storePathType,
-          dispatch
-        )
-      }))
-    }
-  }, [
-    dataInputState.artifacts,
-    dataInputState.artifactsReferences,
-    dataInputState.featureVectors,
-    dataInputState.featureVectorsReferences,
-    dataInputState.inputProjectItemPathEntered,
-    dataInputState.inputProjectItemReferencePathEntered,
-    dataInputState.inputProjectPathEntered,
-    dataInputState.inputStorePathTypeEntered,
-    dataInputState.project,
-    dataInputState.projectItem,
-    dataInputState.projectItemReference,
-    dataInputState.projects,
-    dataInputState.storePathType,
-    dispatch,
-    formState.values.target_path.fieldInfo.pathType,
-    setDataInputState
-  ])
-
-  useEffect(() => {
-    if (
-      dataInputState.inputProjectPathEntered &&
-      dataInputState.storePathType &&
-      dataInputState.project
-    ) {
-      if (dataInputState.storePathType === 'artifacts' && dataInputState.artifacts.length === 0) {
-        dispatch(artifactsAction.fetchArtifacts(dataInputState.project)).then(artifacts => {
-          setDataInputState(prev => ({
-            ...prev,
-            artifacts: generateArtifactsList(artifacts)
-          }))
-        })
-      } else if (
-        dataInputState.storePathType === 'feature-vectors' &&
-        dataInputState.featureVectors.length === 0
-      ) {
-        dispatch(featureStoreActions.fetchFeatureVectors(dataInputState.project)).then(
-          featureVectors => {
-            const featureVectorsList = uniqBy(featureVectors, 'metadata.name')
-              .map(featureVector => ({
-                label: featureVector.metadata.name,
-                id: featureVector.metadata.name
-              }))
-              .sort((prevFeatureVector, nextFeatureVector) =>
-                prevFeatureVector.id.localeCompare(nextFeatureVector.id)
-              )
-
-            setDataInputState(prev => ({
-              ...prev,
-              featureVectors: featureVectorsList
-            }))
-          }
-        )
-      }
-    }
-  }, [
-    dataInputState.artifacts.length,
-    dataInputState.featureVectors.length,
-    dataInputState.inputProjectPathEntered,
-    dataInputState.project,
-    dataInputState.storePathType,
-    dispatch,
-    setDataInputState
-  ])
-
-  useEffect(() => {
-    const storePathType = dataInputState.storePathType
-    const projectName = dataInputState.project
-    const projectItem = dataInputState.projectItem
-
-    if (dataInputState.inputProjectItemPathEntered && storePathType && projectName && projectItem) {
-      if (storePathType === 'artifacts' && dataInputState.artifactsReferences.length === 0) {
-        dispatch(artifactsAction.fetchArtifact(projectName, projectItem)).then(artifacts => {
-          if (artifacts.length > 0 && artifacts[0].data) {
-            setDataInputState(prev => ({
-              ...prev,
-              artifactsReferences: generateArtifactsReferencesList(artifacts[0].data)
-            }))
-          }
-        })
-      } else if (
-        storePathType === 'feature-vectors' &&
-        dataInputState.featureVectorsReferences.length === 0
-      ) {
-        dispatch(featureStoreActions.fetchFeatureVector(projectName, projectItem)).then(
-          featureVectors => {
-            const featureVectorsReferencesList = featureVectors
-              .map(featureVector => {
-                let featureVectorReference = getFeatureReference(featureVector.metadata)
-
-                return {
-                  label: featureVectorReference,
-                  id: featureVectorReference,
-                  customDelimiter: featureVectorReference[0]
-                }
-              })
-              .filter(featureVector => featureVector.label !== '')
-              .sort((prevRef, nextRef) => prevRef.id.localeCompare(nextRef.id))
-
-            setDataInputState(prev => ({
-              ...prev,
-              featureVectorsReferences: featureVectorsReferencesList
-            }))
-          }
-        )
-      }
-    }
-  }, [
-    dataInputState.artifactsReferences.length,
-    dataInputState.featureVectorsReferences.length,
-    dataInputState.inputProjectItemPathEntered,
-    dataInputState.project,
-    dataInputState.projectItem,
-    dataInputState.storePathType,
-    dispatch,
-    setDataInputState
-  ])
 
   return (
     <div className="form">
@@ -285,51 +88,21 @@ const RegisterArtifactModalForm = ({ formState, showType, messageByKind, setFiel
         )}
       </div>
       <div className="form-row">
-        <FormCombobox
-          density="dense"
-          hideSearchInput={!dataInputState.inputStorePathTypeEntered}
-          inputPlaceholder={
-            pathPlaceholders[formState.values.target_path.fieldInfo?.pathType] ?? ''
-          }
-          invalidText={`Field must be in "${
-            pathTips(dataInputState.storePathType)[formState.values.target_path.fieldInfo?.pathType]
-          }" format`}
-          maxSuggestedMatches={
-            formState.values.target_path.fieldInfo.pathType === MLRUN_STORAGE_INPUT_PATH_SCHEME
-              ? 3
-              : 2
-          }
-          name="target_path.path"
-          onChange={(selectValue, inputValue) => handleOnChange(selectValue, inputValue)}
-          required
-          selectOptions={comboboxSelectList}
-          selectPlaceholder="Path Scheme"
-          suggestionList={
-            formState.values.target_path.fieldInfo.pathType === MLRUN_STORAGE_INPUT_PATH_SCHEME
-              ? dataInputState.comboboxMatches
-              : []
-          }
-          validator={(fieldValue, allValues) => validatePath(allValues)}
-        />
-      </div>
-      <div className="form-row">
         <FormTextarea label="Description" maxLength={500} name="description" />
       </div>
-      <OnChange name="target_path.path">
-        {value => {
-          if (value.length !== 0) {
-            formState.form.change('target_path.fieldInfo.value', value.replace(/.*:[/]{2,3}/g, ''))
-            formState.form.change('target_path.fieldInfo.pathType', value.match(/^\w*:[/]{2,3}/)[0])
-          }
-        }}
-      </OnChange>
-      <OnChange name="target_path.fieldInfo.pathType">
-        {(value, prevValue) => {
-          if (prevValue === MLRUN_STORAGE_INPUT_PATH_SCHEME) {
-            setDataInputState(dataInputInitialState)
-          }
-        }}
-      </OnChange>
+      <div className="form-row">
+        <TargetPath
+          density="dense"
+          formState={formState}
+          formStateFieldInfo="target_path.fieldInfo"
+          label="Target Path"
+          name="target_path.path"
+          onChangeListenerName="target_path.fieldInfo.pathType"
+          required
+          selectPlaceholder="Path Scheme"
+          setFieldState={setFieldState}
+        />
+      </div>
     </div>
   )
 }
